Move TaskExecutor to own file
